Split out BitBucket tests
