+++ .circleci/config.yml
@@ -49,7 +49,8 @@ workflows:
               testing_profile: ["non-confidential-tests",
               "confidential-workflow-tests",
               "confidential-tool-tests",
-              "singularity-tests"]
+              "singularity-tests",
+              "bitbucket-tests"]
           <<: *common_filters
           requires:
             - build
