+++ dockstore-cli-integration-testing/src/test/java/io/dockstore/client/cli/GeneralWorkflowIT.java
@@ -404,27 +404,6 @@ public class GeneralWorkflowIT extends BaseIT {
                 SourceControl.GITHUB.toString() + "/DockstoreTestUser2/hello-dockstore-workflow:testCWL", "--script" });
     }
 
-    /**
-     * Tests that convert with valid imports will work (for WDL)
-     */
-    @Test
-    public void testRefreshAndConvertWithImportsWDL() {
-        refreshByOrganizationReplacement(USER_2_USERNAME);
-        Client.main(
-            new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "update_workflow", "--entry",
-                SourceControl.BITBUCKET.toString() + "/dockstore_testuser2/dockstore-workflow", "--descriptor-type", "wdl",
-                "--workflow-path", "/Dockstore.wdl", "--default-test-parameter-path", "/foo.json", "--script" });
-
-        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "refresh", "--entry",
-            SourceControl.BITBUCKET.toString() + "/dockstore_testuser2/dockstore-workflow", "--script" });
-        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "publish", "--entry",
-            SourceControl.BITBUCKET.toString() + "/dockstore_testuser2/dockstore-workflow", "--script" });
-
-        Client.main(
-            new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "convert", "entry2json", "--entry",
-                SourceControl.BITBUCKET.toString() + "/dockstore_testuser2/dockstore-workflow:wdl_import", "--script" });
-        assertTrue(systemOutRule.getLog().contains("\"three_step.cgrep.pattern\": \"String\""));
-    }
 
     /**
      * Tests that a developer can launch a CWL workflow locally, instead of getting files from Dockstore
@@ -680,46 +659,6 @@ public class GeneralWorkflowIT extends BaseIT {
     }
 
     /**
-     * This tests the dirty bit attribute for workflow versions with bitbucket
-     */
-    @Test
-    public void testBitbucketDirtyBit() {
-        refreshByOrganizationReplacement(USER_2_USERNAME);
-
-        // refresh individual that is valid
-        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "refresh", "--entry",
-            SourceControl.BITBUCKET.toString() + "/dockstore_testuser2/dockstore-workflow", "--script" });
-        final long nullLastModifiedWorkflowVersions = testingPostgres
-            .runSelectStatement("select count(*) from workflowversion where lastmodified is null", long.class);
-        assertEquals("All Bitbucket workflow versions should have last modified populated after refreshing", 0,
-            nullLastModifiedWorkflowVersions);
-        // Check that no versions have a true dirty bit
-        final long count = testingPostgres.runSelectStatement("select count(*) from workflowversion where dirtybit = true", long.class);
-        assertEquals("there should be no versions with dirty bit, there are " + count, 0, count);
-
-        // Edit workflow path for a version
-        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "version_tag", "--entry",
-            SourceControl.BITBUCKET.toString() + "/dockstore_testuser2/dockstore-workflow", "--name", "master", "--workflow-path",
-            "/Dockstoredirty.cwl", "--script" });
-
-        // There should be on dirty bit
-        final long count1 = testingPostgres.runSelectStatement("select count(*) from workflowversion where dirtybit = true", long.class);
-        assertEquals("there should be 1 versions with dirty bit, there are " + count1, 1, count1);
-
-        // Update default cwl
-        Client.main(
-            new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "update_workflow", "--entry",
-                SourceControl.BITBUCKET.toString() + "/dockstore_testuser2/dockstore-workflow", "--workflow-path", "/Dockstoreclean.cwl",
-                "--script" });
-
-        // There should be 3 versions with new cwl
-        final long count2 = testingPostgres
-            .runSelectStatement("select count(*) from workflowversion where workflowpath = '/Dockstoreclean.cwl'", long.class);
-        assertEquals("there should be 4 versions with workflow path /Dockstoreclean.cwl, there are " + count2, 4, count2);
-
-    }
-
-    /**
      * This is a high level test to ensure that gitlab basics are working for gitlab as a workflow repo
      */
     @Test
@@ -811,31 +750,6 @@ public class GeneralWorkflowIT extends BaseIT {
 
     }
 
-    /**
-     * This tests manually publishing a Bitbucket workflow
-     */
-    @Test
-    public void testManualPublishBitbucket() {
-        // manual publish
-        Client.main(
-            new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "manual_publish", "--repository",
-                "dockstore-workflow", "--organization", "dockstore_testuser2", "--git-version-control", "bitbucket", "--workflow-name",
-                "testname", "--workflow-path", "/Dockstore.wdl", "--descriptor-type", "wdl", "--script" });
-
-        // Check for two valid versions (wdl_import and surprisingly, cwl_import)
-        final long count = testingPostgres
-            .runSelectStatement("select count(*) from workflowversion where valid='t' and (name='wdl_import' OR name='cwl_import')",
-                long.class);
-        assertEquals("There should be a valid 'wdl_import' version and a valid 'cwl_import' version", 2, count);
-
-        final long count2 = testingPostgres
-            .runSelectStatement("select count(*) from workflowversion where lastmodified is null", long.class);
-        assertEquals("All Bitbucket workflow versions should have last modified populated when manual published", 0, count2);
-
-        // grab wdl file
-        Client.main(new String[] { "--config", ResourceHelpers.resourceFilePath("config_file2.txt"), "workflow", "wdl", "--entry",
-            SourceControl.BITBUCKET.toString() + "/dockstore_testuser2/dockstore-workflow/testname:wdl_import", "--script" });
-    }
 
     /**
      * This tests manually publishing a gitlab workflow
