+++ pom.xml
@@ -656,7 +656,8 @@
                     io.dockstore.common.SlowTest,
                     io.dockstore.common.BenchmarkTest,
                     io.dockstore.common.ToilOnlyTest,
-                    io.dockstore.client.cli.nested.WestTests
+                    io.dockstore.client.cli.nested.WestTests,
+                    io.dockstore.common.BitBucketTest
                 </excludeGroups>
             </properties>
 
@@ -711,7 +712,8 @@
                     io.dockstore.client.cli.nested.SingularityTest,
                     io.dockstore.common.SlowTest,
                     io.dockstore.common.ToilOnlyTest,
-                    io.dockstore.client.cli.nested.WesTests
+                    io.dockstore.client.cli.nested.WesTests,
+                    io.dockstore.common.BitBucketTest
                 </excludeGroups>
             </properties>
         </profile>
@@ -727,7 +729,8 @@
                     io.dockstore.common.SlowTest,
                     io.dockstore.common.ToilOnlyTest,
                     io.dockstore.common.WorkflowTest,
-                    io.dockstore.client.cli.nested.WesTests
+                    io.dockstore.client.cli.nested.WesTests,
+                    io.dockstore.common.BitBucketTest
                 </excludeGroups>
             </properties>
         </profile>
@@ -743,8 +746,18 @@
                     io.dockstore.common.SlowTest,
                     io.dockstore.common.ToilOnlyTest,
                     io.dockstore.common.ToolTest,
-                    io.dockstore.client.cli.nested.WesTests
+                    io.dockstore.client.cli.nested.WesTests,
+                    io.dockstore.common.BitBucketTest
                 </excludeGroups>
+            </properties>
+        </profile>
+        <profile>
+            <id>bitbucket-tests</id>
+            <!-- run non-confidential tests -->
+            <properties>
+                <skipTests>true</skipTests>
+                <skipITs>false</skipITs>
+                <Groups>io.dockstore.common.BitBucketTest</Groups>
             </properties>
         </profile>
         <profile>
@@ -758,7 +771,8 @@
                     io.dockstore.common.SlowTest,
                     io.dockstore.common.ToilOnlyTest,
                     io.dockstore.common.ConfidentialTest,
-                    io.dockstore.client.cli.nested.WesTests
+                    io.dockstore.client.cli.nested.WesTests,
+                    io.dockstore.common.BitBucketTest
                 </excludeGroups>
             </properties>
         </profile>
